+++ packages/server/index.js
@@ -1,18 +1,30 @@
-const { initializeStartTime } = require('./lib/util/performance_benchmark')
-
-const run = async () => {
-  initializeStartTime()
-
-  if (require.name !== 'customRequire') {
-    // Purposefully make this a dynamic require so that it doesn't have the potential to get picked up by snapshotting mechanism
-    const hook = './hook'
-
-    const { hookRequire } = require(`${hook}-require`)
-
-    hookRequire(false)
-  }
-
-  await require('./server-entry')
+const runChildProcess = async (entryPoint) => {
+  require('./lib/plugins/child/register_ts_node')
+  require(entryPoint)
 }
 
-module.exports = run()
+const startCypress = async () => {
+  try {
+    const { initializeStartTime } = require('./lib/util/performance_benchmark')
+
+    initializeStartTime()
+
+    const { hookRequire } = require('./hook-require')
+
+    hookRequire({ forceTypeScript: false })
+
+    await require('./start-cypress')
+  } catch (error) {
+    // eslint-disable-next-line no-console
+    console.error(error)
+    process.exit(1)
+  }
+}
+
+const { entryPoint } = require('minimist')(process.argv.slice(1))
+
+if (entryPoint) {
+  module.exports = runChildProcess(entryPoint)
+} else {
+  module.exports = startCypress()
+}